Added DDP Example to Torchsnapshot
Summary:
We added  an example to torchsnapshot/examples that demonstrates TorchSnapshot integration with DistribututedDataParallel (DDP). Similar to other existing examples, it will help users get started more easily with TorchSnapshot.

The new example, torchsnapshot/examples/ddp_example.py, follows the structure in the single-process example and includes
a DDP wrapper around the model and
a torch.distributed.init_process_group call. We also configured it to run with TorchElastic.

Reviewed By: ananthsub, yifuwang

Differential Revision: D37425291

fbshipit-source-id: acb6d699dae0b28d577382ec3b3123706850b8d6