Implement chunking for replicated tensors in torchsnapshot.
Summary:
Implemented ChunkedTensorIOPreparer class, which subdivide replicated tensors into chunks in order to return a ChunkedTensorEntry object and corresponding write requests.
Moreover, user can elect whether or not to enable chunking (which they can do by directly setting TORCHSNAPSHOT_ENABLE_CHUNKING=True). From a developer's perspective, this means that this implementation is lower-risk to land, as the base code for torchsnapshot (which does not enable chunking) is totally intact.

Reviewed By: yifuwang

Differential Revision: D38295461

fbshipit-source-id: 0e4f38855272b000cfd21a0ea16658c46892c0ad