Added test to verify that when the metadata is serialized via json, it is load-able with the yaml loader.
Summary:
When the number of entries in the snapshot metadata is very large, yaml
serialization becomes a bottleneck and there's little we can do to
optimize. We likely need to switch to json to overcome this. Fortunately,
when our metadata is serialized via json, it is compatible with the yaml
loader, so we can make the switch in a backward compatible fashion. This
test makes sure that we don't do anything crazy with the metadata to break
this compatibility.

Reviewed By: edward-io

Differential Revision: D40207674

fbshipit-source-id: 54c96fb3e0d5126a0f0badf497680736798eadcf