reconstruct the original state dict on load (#104)
Summary:
Pull Request resolved: https://github.com/pytorch/torchsnapshot/pull/104

Previous, on `.restore()`, we retrieve the state dict of the target stateful, restore its values in-place, and use the state dict to invoke the target stateful's `.load_state_dict()`.

This approach has a flaw: sometimes it's valid for the persisted state dict to have a different structure compared to the target stateful's state dict, in which case the correct behavior is to reconstruct the persisted state dict.

This diff changes the following:
- `get_available_entry()` now returns the container entries for reconstructing the persisted state dict
- `.load_stateful()` would still in-place load tensors that are found in the target stateful's state dict. But it would also load tensors that are not found in the target stateful's state dict
- The state dict is now inflated with the persisted container entries as opposed to the target stateful's container entries

Reviewed By: ananthsub

Differential Revision: D40462737

fbshipit-source-id: 2869e6a4c3d33331a129ffd56557ae6b41b914a5