Default sharded tensor elasticity check (#154)
Summary:
Pull Request resolved: https://github.com/pytorch/torchsnapshot/pull/154

Currently for sharded tensor elasticity, torchsnapshot only searches for sharded tensors at the root of the state dict. As the note in the comment indicates, this means elasticity for most optimizers is not supported. This change aims to lift this restriction.

This change in default behavior can be toggled off by setting `TORCHSNAPSHOT_ENABLE_SHARDED_TENSOR_ELASTICITY_ROOT_ONLY=1`

Reviewed By: daniellepintz

Differential Revision: D49620100

fbshipit-source-id: 47f791f3805dc9c4bdf8778e582aa9908d2e6d61