Handle replicated paths for DTensor
Summary: Storage paths/logical paths for entries can begin with `sharded/` or `replicated/` for either of those cases. DTensors can be both sharded and replicated. This diff updates `get_storage_path` for this case (I simply used `replicated_sharded/` as the prefix, open to other suggestions) and adds some utilities for checking if a tensor/dtensor is sharded or replicated. Wasn't sure where to put this util so I placed it in a new `utils.py` file, also open to suggestions here.

Reviewed By: ananthsub

Differential Revision: D49176000

fbshipit-source-id: 924e174d6fcaa01bd058ccf36c3c2b1593586dec