Pass in `strict` flag to allow non-strict state_dict loading
Summary:
X-link: https://github.com/pytorch/tnt/pull/691

This allows users to ignore keys in the state_dict that aren't part of
the given module, or part of the module that aren't in the state_dict.

The default value for the flag (true) keeps the status quo and what the pytorch
interface uses.

Reviewed By: anshulverma

Differential Revision: D53066198

fbshipit-source-id: 8a849f46d09d6e7d9185d589b966f7a7a089d9fc