For each checkpoint attempt, only emit 1 signpost event
Summary:
Currently, as each rank attempts to log its event to signpost, we end up with multiple signpost logs for the same checkpoint attempt. For examples https://fburl.com/scuba/workflow_signpost/ssez2lok. This affects the scalability of the MLU consumption code to calculate MLU. We want to reduce to single signpost per checkpoint.

This diff makes the change for the signpost logging so it happens on rank 0

Reviewed By: galrotem

Differential Revision: D54345169

fbshipit-source-id: 63789a3024c902b7ec6272fee1e0a814040be201