Bring back a limited nest_asyncio
Summary:
This brings back a limited subset of `nest_asyncio` to allow re-entrant calls into the event loop, but avoid monkey-patching the global namespace.

I don't really like this solution, but it seems the cleanest way to unblock suing TorchSnapshot in Bento for now.
A cleaner solution would be to refactor the code to *stop* passing around an event loop and instead rely on `nested_asyncio_run(Coroutine[T]) -> T` function (which we can implement relatively easily), but that seemed to be a much larger refactor.

Reviewed By: JKSenthil

Differential Revision: D60845357

fbshipit-source-id: 3873e507ddbfce219233aea4c759c55e0f574bd5